First pass at introducing slog into cernan
This commit is a very preliminary introduction of slog into cernan,
replacing the old 'log' module. There's a fair bit of work that needs
to be done to make this happen. The end-goal is to have cernan kick
out structured logs about its operation. Once we've got that we can
do all sorts of nifty integration testing work.

Signed-off-by: Brian L. Troutwine <brian@troutwine.us>