Further work, now compiling
This is... a hot mess. I figured it would be. There's been automatic
rustfmt passes on the majority of the codebase at ths point as I've
touched most of the codebase. That's okay though.

At present we don't have the same logging facility that we once enjoyed.
I need to figure out how to set up drains etc to have the old information
that fern gave us.

I don't even know what happens if we run under test.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>