Pin down the shutdown hang issue
I've added a 'shutdown' to the sink to allow for the controlled
shutdown of internal resources. The shutdown will hang for prometheus
sink because of the issue linked in the comments.

We're probably going to have to update prometheus sink HTTP server.
This also implies that, going forward, we'll have to be very careful
about testing shutdown functions for fully running sinks.

Signed-off-by: Brian L. Troutwine <blt@postmates.com>